lockf implementation proposal

+++ src/c/fcntl.c
@@ -14,13 +14,13 @@ int open(const char* filename, int flags, ...) {
     return sys_open(filename, flags, mode);
 }
 
-int sys_fcntl(int fildes, int cmd, int args);
+int sys_fcntl(int fildes, int cmd, unsigned long long args);
 
 int fcntl(int fildes, int cmd, ...) {
-    int args = 0;
+    unsigned long long args = 0;
     va_list ap;
     va_start(ap, cmd);
-    args = va_arg(ap, int);
+    args = va_arg(ap, unsigned long long);
     va_end(ap);
     return sys_fcntl(fildes, cmd, args);
 }
